# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Solunex Lab Backend",
                 "version":  "0.1.0"
             },
    "paths":  {
...
                  "/api/patients":  {
                                        "post":  {
                                                     "tags":  [
                                                                  "patients"
                                                              ],
                                                     "summary":  "Create Patient",
                                                     "operationId":  "create_patient_api_patients_post",
                                                     "requestBody":  {
                                                                         "content":  {
                                                                                         "application/json":  {
                                                                                                                  "schema":  {
                                                                                                                                 "$ref":  "#/components/schemas/PatientCreate"
                                                                                                                             }
                                                                                                              }
                                                                                     },
                                                                         "required":  true
                                                                     },
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "Successful Response",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/PatientOut"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               },
                                                                       "422":  {
                                                                                   "description":  "Validation Error",
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                       }
                                                                                                                        }
                                                                                               }
                                                                               }
                                                                   }
                                                 }
                                    },
                  "/api/patients/search":  {
                                               "get":  {
                                                           "tags":  [
                                                                        "patients"
                                                                    ],
                                                           "summary":  "Search Patients",
                                                           "operationId":  "search_patients_api_patients_search_get",
                                                           "parameters":  [
                                                                              {
                                                                                  "name":  "q",
                                                                                  "in":  "query",
                                                                                  "required":  true,
                                                                                  "schema":  {
                                                                                                 "type":  "string",
                                                                                                 "minLength":  1,
                                                                                                 "title":  "Q"
                                                                                             }
                                                                              }
                                                                          ],
                                                           "responses":  {
                                                                             "200":  {
                                                                                         "description":  "Successful Response",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "type":  "array",
                                                                                                                                                 "items":  {
                                                                                                                                                               "$ref":  "#/components/schemas/PatientOut"
                                                                                                                                                           },
                                                                                                                                                 "title":  "Response Search Patients Api Patients Search Get"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     },
                                                                             "422":  {
                                                                                         "description":  "Validation Error",
                                                                                         "content":  {
                                                                                                         "application/json":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                             }
                                                                                                                              }
                                                                                                     }
                                                                                     }
                                                                         }
                                                       }
                                           },
                  "/api/patients/{patient_id}":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "patients"
                                                                          ],
                                                                 "summary":  "Get Patient",
                                                                 "operationId":  "get_patient_api_patients__patient_id__get",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "patient_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "Patient Id"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/PatientOut"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             },
                                                     "patch":  {
                                                                   "tags":  [
                                                                                "patients"
                                                                            ],
                                                                   "summary":  "Update Patient",
                                                                   "operationId":  "update_patient_api_patients__patient_id__patch",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "patient_id",
                                                                                          "in":  "path",
                                                                                          "required":  true,
                                                                                          "schema":  {
                                                                                                         "type":  "integer",
                                                                                                         "title":  "Patient Id"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "requestBody":  {
                                                                                       "required":  true,
                                                                                       "content":  {
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/PatientUpdate"
                                                                                                                                           }
                                                                                                                            }
                                                                                                   }
                                                                                   },
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Successful Response",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/PatientOut"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             },
                                                                                     "422":  {
                                                                                                 "description":  "Validation Error",
                                                                                                 "content":  {
                                                                                                                 "application/json":  {
                                                                                                                                          "schema":  {
                                                                                                                                                         "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                     }
                                                                                                                                      }
                                                                                                             }
                                                                                             }
                                                                                 }
                                                               }
                                                 },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
                                       "PatientCreate":  {
                                                             "properties":  {
                                                                                "patient_no":  {
                                                                                                   "type":  "string",
                                                                                                   "maxLength":  50,
                                                                                                   "minLength":  1,
                                                                                                   "title":  "Patient No"
                                                                                               },
                                                                                "full_name":  {
                                                                                                  "type":  "string",
                                                                                                  "maxLength":  255,
                                                                                                  "minLength":  1,
                                                                                                  "title":  "Full Name"
                                                                                              },
                                                                                "phone":  {
                                                                                              "type":  "string",
                                                                                              "maxLength":  30,
                                                                                              "minLength":  3,
                                                                                              "title":  "Phone"
                                                                                          },
                                                                                "date_of_birth":  {
                                                                                                      "anyOf":  [
                                                                                                                    {
                                                                                                                        "type":  "string",
                                                                                                                        "format":  "date"
                                                                                                                    },
                                                                                                                    {
                                                                                                                        "type":  "null"
                                                                                                                    }
                                                                                                                ],
                                                                                                      "title":  "Date Of Birth"
                                                                                                  },
                                                                                "gender":  {
                                                                                               "anyOf":  [
                                                                                                             {
                                                                                                                 "type":  "string"
                                                                                                             },
                                                                                                             {
                                                                                                                 "type":  "null"
                                                                                                             }
                                                                                                         ],
                                                                                               "title":  "Gender"
                                                                                           },
                                                                                "address":  {
                                                                                                "anyOf":  [
                                                                                                              {
                                                                                                                  "type":  "string"
                                                                                                              },
                                                                                                              {
                                                                                                                  "type":  "null"
                                                                                                              }
                                                                                                          ],
                                                                                                "title":  "Address"
                                                                                            }
                                                                            },
                                                             "type":  "object",
                                                             "required":  [
                                                                              "patient_no",
                                                                              "full_name",
                                                                              "phone"
                                                                          ],
                                                             "title":  "PatientCreate"
                                                         },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           }
                                   }
                   }
}